requirements.txt
Django>=4.2
django-environ==0.11.2
django-cors-headers
djangorestframework
psycopg2-binary
drf-spectacular==0.27.0
django-storages==1.14.2
boto3==1.34.34
Pillow==10.2.0
stripe
structlog
